Wanderer: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@39258 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/system/Wanderer/Tools/DiskInfo/support.c
@@ -93,9 +93,9 @@ ULONG FormatSize(STRPTR buffer, ULONG blocks, ULONG totalblocks, ULONG bytesperb
     }
     
     if (!showPercentage)
-        sprintf(buffer, "%.1f%s  (%d %s)", internalsize, suffixes[divcount], blocks, _(MSG_BLOCKS) );
+        sprintf(buffer, "%.1f%s  (%d %s)", internalsize, suffixes[divcount], (int)blocks, _(MSG_BLOCKS) );
     else
-        sprintf(buffer, "%.1f%s  (%d %s, %d%%)", internalsize, suffixes[divcount], blocks, _(MSG_BLOCKS), percentage);
+        sprintf(buffer, "%.1f%s  (%d %s, %d%%)", internalsize, suffixes[divcount], (int)blocks, _(MSG_BLOCKS), (int)percentage);
     
     return percentage;
 }
